Adds support for page children.
Curently only works for static pages with subpages to display. Only one level
deep. Links to subpages, with a link to the top-level parent, are displayed in a
smaller right-hand column at widths greater than 800px.

Does not include these sub pages in the top navigation. (That is a challenge for
another day.)

Fixes #28.